feat: 向下封装

+++ src/sock.h
@@ -2,7 +2,7 @@
 struct sock;
 
 struct sock {
-    int protocol;
+    uint8_t protocol;
     uint16_t sport;
     uint16_t dport;
     uint32_t saddr;
